app/web/requirements.txt
pip
reflex
bcrypt
sqlalchemy
psycopg2
pyjwt
reflex==0.8.3